Auto-generate CMakeLists.txt for project folder

+++ CodeJam/createFolders.bash
@@ -6,13 +6,26 @@ PROBLEM_FOLDERS=("${@:2}")
 IFS=$'\n'
 DIR="$( cd "$( dirname $0 )" && pwd -P )"
 
+if [[ $TARGET_DIR = "" ]]
+then
+    echo You must provide 1 argument and optional folder arguments:
+    echo "  First: directory location"
+    echo "  Optional: letters/names of problem folders to create!"
+    exit 1
+fi
+
 rm -rf $DIR/template_problem_folder
 
+TARGET_ABSOLUTE=$(readlink -f "$TARGET_DIR")
+TARGET_BASENAME=$(basename "$TARGET_ABSOLUTE")
+
 bash $DIR/mainCreation.bash
+bash $DIR/createCMakeLists.bash $TARGET_BASENAME ${PROBLEM_FOLDERS[@]}
 
 mkdir $DIR/template_problem_folder
 cp $DIR/Main.cpp $DIR/template_problem_folder/Main.cpp
 cp $DIR/Solution.py.m4 $DIR/template_problem_folder/Solution.py.m4
+cp $DIR/CMakeLists.txt $DIR/template_problem_folder/CMakeLists.txt
 touch $DIR/template_problem_folder/sample.txt
 touch $DIR/template_problem_folder/result.txt
 cp $DIR/interactive_runner.py $DIR/template_problem_folder/interactive_runner.py
@@ -20,13 +33,6 @@ cp $DIR/execute.bash $DIR/template_problem_folder/execute.bash
 cp $DIR/sync_lines_after_m4.py $DIR/template_problem_folder/sync_lines_after_m4.py
 
 
-if [[ $TARGET_DIR = "" ]]
-then
-    echo You must provide 1 argument and optional folder arguments:
-    echo "  First: directory location"
-    echo "  Optional: letters/names of problem folders to create!"
-    exit
-fi
 
 mkdir -p "$TARGET_DIR"
 cp -r $DIR/template_problem_folder/* "$TARGET_DIR"
